// src/app/core/services/model/projectBaseData.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProjectPersonMetaData } from './projectPersonMetaData';

export interface ProjectBaseData { 
    name: string;
    place: string;
    postalCode: string;
    street: string;
    imageUrl?: string;
    projectResponsible: ProjectPersonMetaData;
    projectInstalledBy: ProjectPersonMetaData;
}